# openapi/wms-api.yaml
openapi: 3.1.0
info:
  title: WMS ShowCase API
  description: 倉庫管理システム（WMS）ShowCase のREST API定義
  version: 0.1.0
  contact:
    name: WMS ShowCase Project
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://{environment}.azurecontainerapps.io
    description: Azure Container Apps
    variables:
      environment:
        default: wms-showcase
...
paths:
...
  /api/v1/master/warehouses:
    get:
      operationId: listWarehouses
      summary: 倉庫一覧取得
      description: 倉庫マスタの一覧を取得する。`all=true`でプルダウン用の全件取得も可能
      tags: [master-warehouse]
      parameters:
        - name: warehouseCode
          in: query
          schema:
            type: string
            maxLength: 50
          description: 倉庫コード（前方一致）
        - name: warehouseName
          in: query
          schema:
            type: string
            maxLength: 200
          description: 倉庫名（部分一致）
        - name: isActive
          in: query
          schema:
            type: boolean
          description: 有効/無効フィルタ。省略時は全件
        - name: all
          in: query
          schema:
            type: boolean
          description: trueの場合ページングなしで全件返却（プルダウン用）
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: sort
          in: query
          schema:
            type: string
            default: warehouseCode,asc
          description: ソート条件
      responses:
        '200':
          description: 倉庫一覧取得成功
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/WarehousePageResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/WarehouseSimple'
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createWarehouse
      summary: 倉庫登録
      description: 倉庫マスタに新規倉庫を登録する
      tags: [master-warehouse]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWarehouseRequest'
      responses:
        '201':
          description: 倉庫登録成功
          headers:
            Location:
              description: 作成されたリソースのURL
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 倉庫コード重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateCode:
                  value:
                    errorCode: DUPLICATE_CODE
                    message: 指定された倉庫コードは既に使用されています
  /api/v1/master/warehouses/{id}:
    get:
      operationId: getWarehouse
      summary: 倉庫取得
      description: 指定IDの倉庫マスタを1件取得する
      tags: [master-warehouse]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 倉庫取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 倉庫が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    errorCode: WAREHOUSE_NOT_FOUND
                    message: 指定された倉庫が見つかりません
    put:
      operationId: updateWarehouse
      summary: 倉庫更新
      description: 指定IDの倉庫マスタを更新する。倉庫コードは変更不可
      tags: [master-warehouse]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWarehouseRequest'
      responses:
        '200':
          description: 倉庫更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 倉庫が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
  /api/v1/master/warehouses/{id}/toggle-active:
    patch:
      operationId: toggleWarehouseActive
      summary: 倉庫無効化／有効化
      description: 指定IDの倉庫の有効/無効を切り替える。在庫が存在する倉庫は無効化不可
      tags: [master-warehouse]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleActiveRequest'
      responses:
        '200':
          description: 無効化/有効化成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseToggleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 倉庫が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
        '422':
          description: 在庫が存在するため無効化不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                hasInventory:
                  value:
                    errorCode: CANNOT_DEACTIVATE_HAS_INVENTORY
                    message: 在庫が存在するため無効化できません
  /api/v1/master/warehouses/exists:
    get:
      operationId: checkWarehouseCodeExists
      summary: 倉庫コード存在確認
      description: 指定された倉庫コードがすでに登録されているか確認する
      tags: [master-warehouse]
      parameters:
        - name: warehouseCode
          in: query
          required: true
          schema:
            type: string
          description: チェック対象の倉庫コード
      responses:
        '200':
          description: 存在確認結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExistsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --- Buildings ---
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWTアクセストークン（httpOnly Cookie）

  # ============================================================
  # Schemas
  # ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
          description: エラーコード
          example: VALIDATION_ERROR
        message:
          type: string
          description: エラーメッセージ
          example: 入力内容に誤りがあります
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          description: フィールドレベルのバリデーションエラー詳細（VALIDATION_ERRORの場合のみ）
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: エラーが発生したフィールド名
          example: userCode
        message:
          type: string
          description: フィールドのエラーメッセージ
          example: 必須項目です
...
    ExistsResponse:
      type: object
      required: [exists]
      properties:
        exists:
          type: boolean
          description: 存在する場合true
...
    ToggleActiveRequest:
      type: object
      required: [isActive, version]
      properties:
        isActive:
          type: boolean
          description: falseで無効化、trueで有効化
        version:
          type: integer
          description: 楽観的ロックバージョン

    # --- Facility Master: Warehouse ---
    CreateWarehouseRequest:
      type: object
      required: [warehouseCode, warehouseName]
      properties:
        warehouseCode:
          type: string
          pattern: '^[A-Z]{4}$'
          description: 倉庫コード（英大文字4文字固定・登録後変更不可）
          example: WARA
        warehouseName:
          type: string
          maxLength: 200
          description: 倉庫名
          example: 東京DC
        warehouseNameKana:
          type: string
          maxLength: 200
          description: 倉庫名カナ
          example: トウキョウディーシー
        address:
          type: string
          maxLength: 500
          description: 住所
          example: 東京都江東区辰巳1-1-1
    UpdateWarehouseRequest:
      type: object
      required: [warehouseName, version]
      properties:
        warehouseName:
          type: string
          maxLength: 200
          description: 倉庫名
        warehouseNameKana:
          type: string
          maxLength: 200
          description: 倉庫名カナ
        address:
          type: string
          maxLength: 500
          description: 住所
        version:
          type: integer
          description: 楽観的ロックバージョン
    WarehouseDetail:
      type: object
      required: [id, warehouseCode, warehouseName, isActive, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        warehouseCode:
          type: string
        warehouseName:
          type: string
        warehouseNameKana:
          type: string
        address:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WarehouseSimple:
      type: object
      required: [id, warehouseCode, warehouseName, isActive]
      properties:
        id:
          type: integer
          format: int64
        warehouseCode:
          type: string
        warehouseName:
          type: string
        isActive:
          type: boolean
    WarehousePageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WarehouseListItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    WarehouseListItem:
      type: object
      required: [id, warehouseCode, warehouseName, isActive, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        warehouseCode:
          type: string
        warehouseName:
          type: string
        warehouseNameKana:
          type: string
        address:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WarehouseToggleResponse:
      type: object
      required: [id, warehouseCode, warehouseName, isActive, version, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        warehouseCode:
          type: string
        warehouseName:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time

    # --- Facility Master: Building ---
...
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: リソースID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: ページ番号（0始まり）
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: ページサイズ
...
  responses:
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              value:
                errorCode: VALIDATION_ERROR
                message: 入力内容に誤りがあります
                details:
                  - field: userCode
                    message: 必須項目です
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                errorCode: UNAUTHORIZED
                message: 認証が必要です
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                errorCode: FORBIDDEN
                message: この操作を実行する権限がありません
    InvalidParameter:
      description: クエリパラメータ不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidParameter:
              value:
                errorCode: INVALID_PARAMETER
                message: クエリパラメータが不正です
    OptimisticLockConflict:
      description: 楽観的ロック競合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              value:
                errorCode: OPTIMISTIC_LOCK_CONFLICT
                message: 他のユーザーによる更新が先行しました。最新データを取得してから再度お試しください
security:
  - cookieAuth: []